BC Auth failure message for unspecified reason code
Some access points/driver implementation send an
association rejection event with the status code
"unspecified" when the password is incorrect.

We now unconditionally broadcast an authentication
failure event(with wrong password error code) to
the framework when such an event is received.
Previously, we used to skip sending the broadcast message
if the network was connected before.

However, it's fine to send the broadcast message with
wrong password error, as the clientModeImpl already has
logic to skip sending the wrong password notification to the user.

By removing the check in the association rejection path,
we can now send a broadcast message about the incorrect password error.
This will allow the clientModeImpl to disable the network
(after receiving three consecutive failures) due to an
incorrect password error, prompting the user to re-enter the password.

This behavior is similar to how the framework handles WPA2 authentication errors.

Bug: 353763640
Test: atest SupplicantStaIfaceHalAidlImplTest SupplicantStaIfaceHalHidlImplTest
Test: Manual - Connect to WPA3-AP. Change the password of AP.
      Confirmed that user gets "check password and try again"
      notification.
Change-Id: I9578f0548cfa9072e52e2c8de52c3a0a211baf66