Fix NPE caused by null value in mCachedStaticChipInfos
HalDeviceManager#mCachedStaticChipInfos should not contain any null
values, but this will happen if we get a JSONException in the middle of
populating its values. Fix this by returning an empty array when we get
a JSONException. The correct values will be populated again when the
driver concurrency combo is loaded after STA bring up.

Flag: EXEMPT bugfix

Bug: 372474058
Test: atest HalDeviceManagerTest
Change-Id: I321a0033489dd14a7557334913bf4eca3dc300a5