Add ART profile to wifi service.
This CL adds an ART profile to the wifi service. This profile will be used to improve the performance of the wifi service. The profile was generated by profiling the wifi service on Droidfood population.

The ART profile is expected to reduce the size of the Wifi service generated code size by 10.82MB! (15.8%)

The new behavior is guarded by a new build feature `RELEASE_SERVICE_WIFI_SPEED_PROFILE_ART_COMPILATION`

Bug: 241823638
Change-Id: I0298cf932bcd39e3b33cbc45df5d8556ba417001
Test: `m`
Test: with flag on: go/art-benchmark?p=BootImageProfileId:41874 (code size reduction by -10.82MB)
Test: with flag off: go/art-benchmark?p=BootImageProfileId:41898 (no code change or metric changes)