Move APIs to check USD support to WifiManager
The current method for checking USD publisher/subscriber support relies
on the UsdManager API, which assumes the USD service is always active.
However, on devices with limited memory where USD functionality might
not be essential, the USD service can be disabled. To address this, it's
proposed to relocate the USD support check APIs to the WiFiManager. This
change would allow for USD support verification even when the USD
service isn't running. This approach provides greater flexibility and
efficiency for platforms where USD services might be optional.

Flag: android.net.wifi.flags.usd
Bug: 340878198
Test: atest FrameworksWifiApiTest
Change-Id: I073c1ba8ea8098c50b805d5a3ac9175fadef6ab8