Update disabling state earlier
Before this fix, when wifi gets disabled, a command is posted to the
ConcreteClientModeManager statemachine. There is a delay to this and
other wifi framework messages could get processed in between, causing
race conditions to happen and break wifi.

Need to update the wifi state to 'disabling' in the same message that
processes the cmd turning wifi off. This is needed so that the
ActiveModeWarden knows wifi is turning off and won't give out additional
secondary STAs in the process.

Flag: EXEMPT bugfix
Bug: 383207702
Test: atest com.android.server.wifi
Test: manually verified
Change-Id: I9223ba11094939d7ce220e047e81f33270af9203