wifi: Update P2P device found with WFD R2 information
This CL adds an update to WifiP2pDevice with WFD R2 specific
information.

1. Update the WifiP2pDevice with supported bootstrapping methods
advertised in P2P_DEVICE_FOUND event from supplicant. This helps
application to know the bootstrapping methods and helps to choose
a pairing bootstrapping method in  group formation or connection.
2. Save the received Device Identity Resolution resolution in
WifiP2pDevice. This is for framework to check if the device is a
previously paired device.

Bug: 341971059
Flag: com.android.wifi.flags.wifi_direct_r2
Test: Manual - Basic P2P connect/disconnect tests
Test: TH Presubmit tests
Test: atest com.android.server.wifi.p2p
Change-Id: I36e3de06cb6c70873443d45b1ec8b857f4facc70